fix reviews
